+++ @navikt/core/react/src/loader/loader.stories.tsx
@@ -7,7 +7,7 @@ export default {
   title: "ds-react/Loader",
   component: Loader,
   parameters: {
-    chromatic: { disable: false },
+    chromatic: { disable: true },
   },
 } satisfies Meta<typeof Loader>;